requirements.txt
# API Framework
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0

# LangChain and Vector Store
langchain>=0.0.267
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
faiss-cpu>=1.7.4
ollama>=0.1.0

# Document Processing
pypdf>=3.15.1
python-multipart>=0.0.6
pandas>=2.0.3
python-magic>=0.4.27
filetype>=1.2.0
aiofiles>=23.2.1

# Environment and Utilities
python-dotenv>=1.0.0
uuid>=1.30

# Additional Dependencies for Metadata and File Handling
PyPDF2>=3.0.0